Extract reusable FeedStoreSpecs helper methods into a shared scope
Extract reusable FeedStoreSpecs helper methods into a shared scope so it can be used by other `FeedStore` implementation tests
